+++ src/main/java/com/github/andrefbsantos/libpricealarm/PriceHitUpperBoundAlarm.java
@@ -7,7 +7,7 @@ public class PriceHitUpperBoundAlarm extends PriceHitAlarm {
 
 	private static final long serialVersionUID = -7107219115788419337L;
 
-	public PriceHitUpperBoundAlarm(long id, Exchange exchange, Pair pair, long period,
+	public PriceHitUpperBoundAlarm(int id, Exchange exchange, Pair pair, long period,
 			Notify notify, double upperBound) throws UpperBoundSmallerThanLowerBoundException {
 		super(id, exchange, pair, period, notify, upperBound, Double.NEGATIVE_INFINITY);
 	}
